renames sampleData to storyData and moves it up to /src level
